+++ tests/lib/Integration/Auth/ResourceAuthorizerTest.php
@@ -99,7 +99,7 @@ class ResourceAuthorizerTest extends TestCase
     {
         $tag = factory(Tag::class)->create();
 
-        $this->doRead($tag->uuid)->assertStatus(401)->assertJson([
+        $this->doRead($tag)->assertStatus(401)->assertJson([
             'errors' => [
                 [
                     'title' => 'Unauthenticated',
@@ -115,19 +115,19 @@ class ResourceAuthorizerTest extends TestCase
 
         $expected = [
             'type' => 'tags',
-            'id' => $tag->uuid,
+            'id' => $tag->getRouteKey(),
             'attributes' => [
                 'created-at' => $tag->created_at->toAtomString(),
                 'updated-at' => $tag->updated_at->toAtomString(),
                 'name' => $tag->name,
             ],
             'links' => [
-                'self' => "http://localhost/api/v1/tags/{$tag->uuid}",
+                'self' => "http://localhost/api/v1/tags/{$tag->getRouteKey()}",
             ],
         ];
 
         $this->actingAsUser('admin')
-            ->doRead($tag->uuid)
+            ->doRead($tag)
             ->assertStatus(200)
             ->assertExactJson(['data' => $expected]);
     }
@@ -137,7 +137,7 @@ class ResourceAuthorizerTest extends TestCase
         $tag = factory(Tag::class)->create();
         $data = [
             'type' => 'tags',
-            'id' => $tag->uuid,
+            'id' => $tag->getRouteKey(),
             'attributes' => [
                 'name' => 'News',
             ],
@@ -158,7 +158,7 @@ class ResourceAuthorizerTest extends TestCase
         $tag = factory(Tag::class)->create();
         $data = [
             'type' => 'tags',
-            'id' => $tag->uuid,
+            'id' => $tag->getRouteKey(),
             'attributes' => [
                 'name' => 'News',
             ],
@@ -179,7 +179,7 @@ class ResourceAuthorizerTest extends TestCase
         $tag = factory(Tag::class)->create();
         $data = [
             'type' => 'tags',
-            'id' => $tag->uuid,
+            'id' => $tag->getRouteKey(),
             'attributes' => [
                 'name' => 'News',
             ],
@@ -195,7 +195,7 @@ class ResourceAuthorizerTest extends TestCase
     {
         $tag = factory(Tag::class)->create();
 
-        $this->doDelete($tag->uuid)->assertStatus(401)->assertJson([
+        $this->doDelete($tag)->assertStatus(401)->assertJson([
             'errors' => [
                 [
                     'title' => 'Unauthenticated',
@@ -211,7 +211,7 @@ class ResourceAuthorizerTest extends TestCase
     {
         $tag = factory(Tag::class)->create();
 
-        $this->actingAsUser()->doDelete($tag->uuid)->assertStatus(403)->assertJson([
+        $this->actingAsUser()->doDelete($tag)->assertStatus(403)->assertJson([
             'errors' => [
                 [
                     'title' => 'Unauthorized',
@@ -228,7 +228,7 @@ class ResourceAuthorizerTest extends TestCase
         $tag = factory(Tag::class)->create();
 
         $this->actingAsUser('admin')
-            ->doDelete($tag->uuid)
+            ->doDelete($tag)
             ->assertStatus(204);
 
         $this->assertDatabaseMissing('tags', ['id' => $tag->getKey()]);